+++ library/src/androidTest/java/com/twilio/video/VideoTrackTopologyParameterizedTest.java
@@ -6,6 +6,7 @@ import android.support.test.runner.AndroidJUnit4;
 import com.twilio.video.base.BaseMediaTest;
 import com.twilio.video.helper.CallbackHelper;
 import com.twilio.video.util.FakeVideoRenderer;
+import com.twilio.video.util.FrameCountRenderer;
 import com.twilio.video.util.Topology;
 
 import org.junit.Before;
@@ -18,12 +19,15 @@ import java.util.List;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
+import static junit.framework.TestCase.assertFalse;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 @RunWith(Parameterized.class)
 @LargeTest
 public class VideoTrackTopologyParameterizedTest extends BaseMediaTest {
+    private static final int VIDEO_TRACK_TEST_DELAY_MS = 3000;
+
     @Parameterized.Parameters(name = "{0}")
     public static Iterable<Object[]> data() {
         return Arrays.asList(new Object[][]{
@@ -65,10 +69,11 @@ public class VideoTrackTopologyParameterizedTest extends BaseMediaTest {
         assertTrue(mediaListener.onVideoTrackAddedLatch.await(20, TimeUnit.SECONDS));
         List<VideoTrack> videoTracks = participant.getMedia().getVideoTracks();
         assertEquals(1, videoTracks.size());
-        FakeVideoRenderer renderer = new FakeVideoRenderer();
-        videoTracks.get(0).addRenderer(renderer);
+        FrameCountRenderer frameCountRenderer = new FrameCountRenderer();
+        videoTracks.get(0).addRenderer(frameCountRenderer);
         assertEquals(1, videoTracks.get(0).getRenderers().size());
-        videoTracks.get(0).removeRenderer(renderer);
+        assertTrue(frameCountRenderer.waitForFrame(VIDEO_TRACK_TEST_DELAY_MS));
+        videoTracks.get(0).removeRenderer(frameCountRenderer);
         assertEquals(0, videoTracks.get(0).getRenderers().size());
     }
 
@@ -82,13 +87,14 @@ public class VideoTrackTopologyParameterizedTest extends BaseMediaTest {
         assertTrue(mediaListener.onVideoTrackAddedLatch.await(20, TimeUnit.SECONDS));
         List<VideoTrack> videoTracks = participant.getMedia().getVideoTracks();
         assertEquals(1, videoTracks.size());
-        FakeVideoRenderer renderer = new FakeVideoRenderer();
+        FrameCountRenderer frameCountRenderer = new FrameCountRenderer();
         VideoTrack videoTrack = videoTracks.get(0);
 
         actor1RoomListener.onParticipantDisconnectedLatch = new CountDownLatch(1);
         actor2Room.disconnect();
         assertTrue(actor1RoomListener.onParticipantDisconnectedLatch.await(20, TimeUnit.SECONDS));
 
-        videoTrack.addRenderer(renderer);
+        videoTrack.addRenderer(frameCountRenderer);
+        assertFalse(frameCountRenderer.waitForFrame(VIDEO_TRACK_TEST_DELAY_MS));
     }
 }